Implement bot release
